+++ packages/nyris-react-components/src/index.tsx
@@ -1,5 +1,5 @@
-export {default as Preview} from "./components/Preview/Preview";
-export {default as Capture} from "./components/Capture/Capture";
+export { default as Preview } from "./components/Preview/Preview";
+export { default as Capture } from "./components/Capture/Capture";
+export { default as Icon } from "./assets/Icon";
 
-export * from './utils';
-
+export * from "./utils";