+++ packages/nyris-webapp/src/components/PanelResult/PostFilter.tsx
@@ -9,11 +9,11 @@ import { useHistory } from 'react-router-dom';
 import { useAppDispatch, useAppSelector } from 'Store/Store';
 import { ExpandablePanelCustom } from './expandable-panel';
 import { getPanelAttributes, getPanelId } from './refinements';
-import { ReactComponent as CloseIcon } from 'common/assets/icons/close.svg';
 import { useTranslation } from 'react-i18next';
 import { setPostFilter } from 'Store/search/Search';
 import { useFilter } from 'hooks/useFilter';
 import { get } from 'lodash';
+import { Icon } from '@nyris/nyris-react-components';
 
 export type ExpandablePanelProps = CurrentRefinementsProvided & {
   children: React.ReactNode;
@@ -244,7 +244,7 @@ export default function PostFilterPanel({
                 height: '32px',
               }}
             >
-              <CloseIcon color="#2B2C46" />
+              <Icon name="close" color="#2B2C46" />
             </Button>
           </div>
         )}