+++ packages/nyris-webapp/src/components/ProductDetailView.tsx
@@ -1,27 +1,21 @@
 import React, { useEffect, useMemo, useState } from 'react';
-import { Button, Collapse, Grid, Typography } from '@material-ui/core';
+import { Button, Collapse, Grid, Typography, Tooltip } from '@material-ui/core';
 import CloseOutlinedIcon from '@material-ui/icons/CloseOutlined';
-import { ReactComponent as IconOpenLink }  from 'common/assets/icons/Union.svg';
 import KeyboardArrowDownIcon from '@material-ui/icons/KeyboardArrowDown';
+import { Icon } from '@nyris/nyris-react-components';
+import NoImage from '../common/assets/images/no-image.svg';
 import { useMediaQuery } from 'react-responsive';
 import { ImagePreviewCarousel } from './carousel/ImagePreviewCarousel';
-import { ReactComponent as IconSearchImage } from 'common/assets/icons/icon_search_image2.svg';
-import { ReactComponent as IconShare } from 'common/assets/icons/Fill.svg';
-import { ReactComponent as IconDisLike } from 'common/assets/icons/icon_dislike.svg';
-import { ReactComponent as IconLike } from 'common/assets/icons/icon_like.svg';
 import { AppState } from 'types';
 import { useAppSelector } from 'Store/Store';
 import { prepareImageList } from '../helpers/CommonHelper';
 import KeyboardArrowUpIcon from '@material-ui/icons/KeyboardArrowUp';
-import NoImage from '../common/assets/images/no-image.svg';
-import { ReactComponent as Box3dIcon } from 'common/assets/icons/3d.svg';
-import { ReactComponent as CloseIcon } from 'common/assets/icons/close.svg';
 import { useTranslation } from 'react-i18next';
 import ProductAttribute from './ProductAttribute';
 import CadenasWebViewer from './CadenasWebViewer';
 import { makeStyles } from '@material-ui/core/styles';
-import { get, isUndefined } from 'lodash';
-import { ReactComponent as IconSettings } from 'common/assets/icons/settings.svg';
+import { get } from 'lodash';
+import { truncateString } from '../helpers/truncateString';
 
 const useStyles = makeStyles(theme => ({
   buttonStyle3D: {
@@ -51,18 +45,14 @@ function ProductDetailView(props: Props) {
   const {
     dataItem,
     handleClose,
-    handlerFeedback,
-    onHandlerModalShare,
     show3dView = false,
     onSearchImage,
   } = props;
-  const { sku, title } = dataItem;
+  const { sku } = dataItem;
   const isMobile = useMediaQuery({ query: '(max-width: 776px)' });
   const { settings } = useAppSelector<AppState>((state: any) => state);
-  const brand = dataItem[settings.field.productTag];
 
-  const [collapDescription, setCollapDescription] = useState(false);
-  const [feedback, setFeedback] = useState('none');
+  const [collapseDescription, setCollapseDescription] = useState(false);
   const [is3dView, setIs3dView] = useState(show3dView);
   const [dataImageCarousel, setDataImageCarouSel] = useState<any[]>([]);
   const [urlImage, setUrlImage] = useState<string>('');
@@ -71,18 +61,6 @@ function ProductDetailView(props: Props) {
   >();
   const { t } = useTranslation();
   const classes = useStyles(props?.show3dView);
-
-  const ctaLink = get(
-    dataItem,
-    settings.field?.ctaLinkField ? settings.field?.ctaLinkField : 'links.main',
-  );
-
-  const secondaryCTALink = get(
-    dataItem,
-    settings.field?.secondaryCTALinkField
-      ? settings.field?.secondaryCTALinkField
-      : '',
-  );
 
   useEffect(() => {
     if (dataItem) {
@@ -127,7 +105,7 @@ function ProductDetailView(props: Props) {
       return details;
     }
   }, [dataItem, settings.field.productDetails]);
-  const manufacturerNumber = get(dataItem, settings.field.manufacturerNumber);
+
   return (
     <div
       className="box-modal-default"
@@ -212,7 +190,7 @@ function ProductDetailView(props: Props) {
                 }
               }}
             >
-              <IconSearchImage color={'#AAABB5'} />
+              <Icon name="search_image" color={'#AAABB5'} />
             </Button>
           )}
           {dataImageCarousel.length === 0 && (
@@ -261,7 +239,7 @@ function ProductDetailView(props: Props) {
                   setIs3dView(true);
                 }}
               >
-                <Box3dIcon width={16} height={16} color={'#AAABB5'} />
+                <Icon name="box3d" width={16} height={16} color={'#AAABB5'} />
               </div>
             )}
           {is3dView && (
@@ -280,7 +258,7 @@ function ProductDetailView(props: Props) {
                 setIs3dView(false);
               }}
             >
-              <CloseIcon width={16} height={16} color={'#FFF'} />
+              <Icon name="close" color={'#FFF'} />
             </div>
           )}
         </div>
@@ -314,56 +292,6 @@ function ProductDetailView(props: Props) {
             }}
           >
             <div className="box-top">
-              {settings.warehouseVariant && (
-                <div
-                  style={{
-                    display: 'flex',
-                    justifyContent: 'space-between',
-                    flexDirection: 'row',
-                    color: settings.theme.mainTextColor ||'#2B2C46',
-                    marginBottom: 10,
-                    paddingLeft: 16,
-                    paddingRight: 16,
-                    gridGap: 8,
-                  }}
-                >
-                  <Typography
-                    className="text-f12 max-line-1 fw-400"
-                    style={{
-                      color: settings.theme.mainTextColor || '#2B2C46',
-                    }}
-                  >
-                    {sku}
-                  </Typography>
-  
-                  {settings.warehouseVariant &&
-                    !isUndefined(
-                      get(dataItem, settings.field.warehouseStockValue),
-                    ) && (
-                      <Typography
-                        className="text-f12 max-line-1 fw-400"
-                        style={{
-                          color: settings.theme.mainTextColor || '#2B2C46',
-                        }}
-                      >
-                        <span
-                          style={{
-                            color: get(
-                              dataItem,
-                              settings.field.warehouseStockValue,
-                            )
-                              ? '#00C070'
-                              : '#c54545',
-                            fontWeight: 600,
-                          }}
-                        >
-                          {get(dataItem, settings.field.warehouseStockValue) || 0}
-                        </span>
-                      </Typography>
-                    )}
-                </div>
-              )}
-  
               <Grid
                 container
                 justifyContent="space-between"
@@ -378,87 +306,93 @@ function ProductDetailView(props: Props) {
                     width: '100%',
                   }}
                 >
-                  {!settings.warehouseVariant && settings.CTAButtonText && (
-                    <ProductAttribute
-                      title={t('Product name')}
-                      value={title}
-                      backgroundColor={settings.theme.brandFieldBackground}
-                      width={
-                        settings.warehouseVariant
-                          ? { xs: '49%', md: 'fit-content' }
-                          : { xs: '100%', md: 'fit-content' }
-                      }
-                    />
-                  )}
-                  {!settings.warehouseVariant && (
-                    <ProductAttribute
-                      title={settings.itemIdLabel || 'SKU'}
-                      value={sku}
-                      backgroundColor={''}
-                      width={
-                        settings.warehouseVariant
-                          ? { xs: '49%', md: 'fit-content' }
-                          : { xs: '100%', md: 'fit-content' }
-                      }
-                    />
-                  )}
-                  {(brand || settings.brandName) && (
-                    <ProductAttribute
-                      title={t('Brand')}
-                      value={brand || settings.brandName}
-                      backgroundColor={settings.theme.brandFieldBackground}
-                      width={
-                        manufacturerNumber
-                          ? { xs: '49%', md: 'fit-content' }
-                          : { xs: '100%', md: 'fit-content' }
-                      }
-                    />
-                  )}
-                  {manufacturerNumber && (
-                    <ProductAttribute
-                      title={t('Manufacturer Number')}
-                      value={manufacturerNumber}
-                      width={
-                        brand || settings.brandName
-                          ? { xs: '49%', md: 'fit-content' }
-                          : { xs: '100%', md: 'fit-content' }
-                      }
-                    />
-                  )}
-                  {settings.warehouseVariant && (
-                    <>
-                      {settings.field.warehouseNumber && (
+                  <div style={{ width: '100%' }}>
+                    {settings.mainTitle && (
+                      <Tooltip
+                        title={dataItem[settings.mainTitle]}
+                        placement="top"
+                        arrow={true}
+                      >
+                        <Typography
+                          className="text-f16 max-line-1 fw-700"
+                          style={{
+                            color: settings.theme.mainTextColor || '#2B2C46',
+                            fontFamily: 'Source Sans 3',
+                            fontSize: '16px',
+                            lineHeight: '22.78px',
+                            width: 'fit-content',
+                            marginLeft: 8,
+                          }}
+                        >
+                          {truncateString(dataItem[settings.mainTitle], 45)}
+                        </Typography>
+                      </Tooltip>
+                    )}
+                    {settings.secondaryTitle && (
+                      <Tooltip
+                        title={dataItem[settings.secondaryTitle] || ''}
+                        placement="top"
+                        arrow={true}
+                      >
+                        <Typography
+                          className="text-f14 max-line-1 fw-400"
+                          style={{
+                            color: settings.theme.mainTextColor || '#2B2C46',
+                            width: 'fit-content',
+                            marginLeft: 8,
+                            fontSize: 14,
+                          }}
+                        >
+                          {truncateString(
+                            dataItem[settings.secondaryTitle],
+                            isMobile ? 45 : 70,
+                          )}
+                        </Typography>
+                      </Tooltip>
+                    )}
+                  </div>
+                  {settings.attributes?.productAttributes && (
+                    <div className="attribute-container">
+                      {!!get(dataItem, settings.attributes?.attributeOneValue || '') && (
                         <ProductAttribute
-                          title={
-                            get(dataItem, settings.field.warehouseNumber) ||
-                            settings.field.warehouseNumber
-                          }
-                          value={
-                            get(dataItem, settings.field.warehouseNumberValue) ||
-                            'N/A'
-                          }
-                          width={{ xs: '49%', md: 'fit-content' }}
+                          title={settings.attributes?.attributeOneLabelValue}
+                          value={get(dataItem, settings.attributes?.attributeOneValue || '')}
+                          padding={settings.theme.brandFieldPadding || '4px 8px'}
+                          backgroundColor={settings.theme.brandFieldBackground}
+                          isTitleVisible={settings.attributes?.labelsAttributes}
                         />
                       )}
-                      {settings.field.warehouseShelfNumber && (
+                      {!!get(dataItem, settings.attributes?.attributeTwoValue || '') && (
                         <ProductAttribute
-                          title={
-                            get(dataItem, settings.field.warehouseShelfNumber) ||
-                            settings.field.warehouseShelfNumber
-                          }
-                          value={
-                            get(
-                              dataItem,
-                              settings.field.warehouseShelfNumberValue,
-                            ) || 'N/A'
-                          }
-                          width={{ xs: '49%', md: 'fit-content' }}
+                          title={settings.attributes?.attributeTwoLabelValue}
+                          value={get(dataItem, settings.attributes?.attributeTwoValue || '')}
+                          padding={settings.theme.brandFieldPadding || '4px 8px'}
+                          backgroundColor={settings.theme.brandFieldBackground}
+                          isTitleVisible={settings.attributes?.labelsAttributes}
                         />
                       )}
-                    </>
+                      {!!get(dataItem, settings.attributes?.attributeThreeValue || '') && (
+                        <ProductAttribute
+                          title={settings.attributes?.attributeThreeLabelValue}
+                          value={get(dataItem, settings.attributes?.attributeThreeValue || '')}
+                          padding={settings.theme.brandFieldPadding || '4px 8px'}
+                          backgroundColor={settings.theme.brandFieldBackground}
+                          isTitleVisible={settings.attributes?.labelsAttributes}
+                        />
+                      )}
+                      {!!get(dataItem, settings.attributes?.attributeFourValue || '') && (
+                        <ProductAttribute
+                          title={settings.attributes?.attributeFourLabelValue}
+                          value={get(dataItem, settings.attributes?.attributeFourValue || '')}
+                          padding={settings.theme.brandFieldPadding || '4px 8px'}
+                          backgroundColor={settings.theme.brandFieldBackground}
+                          isTitleVisible={settings.attributes?.labelsAttributes}
+                        />
+                      )}
+                    </div>
                   )}
                 </div>
-  
+
                 <Grid
                   item
                   xs={12}
@@ -466,98 +400,130 @@ function ProductDetailView(props: Props) {
                     backgroundColor: '#F3F3F5',
                   }}
                 >
-                  {settings.secondaryCTAButtonText && (
-                    <div
-                      style={{
-                        background: settings.theme.secondaryCTAButtonColor || '#2B2C46',
-                        boxShadow: '0px 4px 4px rgba(0, 0, 0, 0.25)',
-                        borderRadius: 4,
-                        marginTop: 8,
-                        display: 'flex',
-                        justifyContent: 'space-between',
-                        alignItems: 'center',
-                      }}
-                      className="btn-detail-item"
-                    >
+                  <div
+                    style={{
+                      width: '100%',
+                      display: 'flex',
+                      flexDirection: isMobile ? 'column' : 'row',
+                      gap: isMobile ? 0 : '6px',
+                    }}
+                  >
+                    {settings.secondaryCTAButton?.secondaryCTAButton && (
                       <div
                         style={{
+                          background: settings.secondaryCTAButton?.secondaryCTAButtonColor || '#2B2C46',
+                          borderRadius: 2,
+                          marginTop: 8,
+                          flex: 1,
                           display: 'flex',
                           justifyContent: 'space-between',
                           alignItems: 'center',
-                          width: '100%',
-                          padding: '0px 12px',
-                          minHeight: 64,
-                          cursor: secondaryCTALink ? 'pointer' : 'normal',
+                          height: isMobile ? '34px' : '50px',
+                          maxWidth: isMobile ? '100%' : 'calc(50% - 3px)',
                         }}
-                        onClick={() => {
-                          if (secondaryCTALink) {
-                            window.open(`${secondaryCTALink}`, '_blank');
-                          }
-                        }}
+                        className="btn-detail-item"
                       >
-                        <Typography
-                          className="text-f18 fw-700 text-white max-line-2"
-                          align="left"
+                        <div
                           style={{
-                            letterSpacing: '0.55px',
-                            maxWidth: '500px',
-                            paddingRight: '4px',
+                            display: 'flex',
+                            justifyContent: 'space-between',
+                            alignItems: 'center',
+                            width: '100%',
+                            padding: '0px 12px',
+                            minHeight: 34,
+                            cursor: settings.secondaryCTAButton?.secondaryCTALinkField ? 'pointer' : 'normal',
+                          }}
+                          onClick={() => {
+                            if (settings.secondaryCTAButton?.secondaryCTALinkField) {
+                              window.open(`${get(dataItem, settings.secondaryCTAButton?.secondaryCTALinkField)}`, '_blank');
+                            }
                           }}
                         >
-                          {settings.secondaryCTAButtonText}
-                        </Typography>
-                        {secondaryCTALink && <IconSettings color="white" />}
+                          <Tooltip
+                            title={settings.secondaryCTAButton?.secondaryCTAButtonText || ''}
+                            placement="top"
+                            arrow={true}
+                          >
+                            <Typography
+                              className="text-f16 fw-600 max-line-1"
+                              align="left"
+                              style={{
+                                letterSpacing: '0.55px',
+                                maxWidth: '220px',
+                                paddingRight: '4px',
+                                color: settings.secondaryCTAButton.secondaryCTAButtonTextColor || '#FFFFFF',
+                              }}
+                            >
+                              {settings.secondaryCTAButton?.secondaryCTAButtonText}
+                            </Typography>
+                          </Tooltip>
+                          {settings.secondaryCTAButton.secondaryCTAIcon && (
+                            <Icon name="settings" color="white" />
+                          )}
+                        </div>
                       </div>
-                    </div>
-                  )}
-  
-                  <div
-                    style={{
-                      background: settings.theme?.CTAButtonColor || settings.theme?.primaryColor,
-                      boxShadow: '0px 4px 4px rgba(0, 0, 0, 0.25)',
-                      borderRadius: 4,
-                      marginTop: 8,
-                      display: 'flex',
-                      justifyContent: 'space-between',
-                      alignItems: 'center',
-                    }}
-                    className="btn-detail-item"
-                  >
-                    <div
-                      style={{
-                        display: 'flex',
-                        justifyContent: 'space-between',
-                        alignItems: 'center',
-                        width: '100%',
-                        padding: '0px 12px',
-                        minHeight: 64,
-                        cursor: ctaLink ? 'pointer' : 'normal',
-                      }}
-                      onClick={() => {
-                        if (ctaLink) {
-                          window.open(`${ctaLink}`, '_blank');
-                        }
-                      }}
-                    >
-                      <Typography
-                        className="text-f18 fw-700 max-line-2"
-                        align="left"
+                    )}
+                    {settings.CTAButton?.CTAButton && (
+                      <div
                         style={{
-                          color: settings.theme?.CTAButtonTextColor || '#FFFFFF',
-                          letterSpacing: '0.55px',
-                          maxWidth: '500px',
-                          paddingRight: '4px',
+                          background: settings.CTAButton?.CTAButtonColor || settings.theme?.primaryColor,
+                          borderRadius: 2,
+                          marginTop: 8,
+                          flex: 1,
+                          display: 'flex',
+                          justifyContent: 'space-between',
+                          alignItems: 'center',
+                          height: isMobile ? '34px' : '50px',
+                          maxWidth: isMobile ? '100%' : 'calc(50% - 3px)',
                         }}
+                        className="btn-detail-item"
                       >
-                        {settings.CTAButtonText
-                          ? settings.CTAButtonText
-                          : dataItem[settings.field.productName]}
-                      </Typography>
-                      {ctaLink && (
-                        <IconOpenLink fill={settings.theme?.CTAButtonTextColor || '#FFFFFF'} width={16} />
-                      )}
-                    </div>
+                        <div
+                          style={{
+                            display: 'flex',
+                            justifyContent: 'space-between',
+                            alignItems: 'center',
+                            width: '100%',
+                            padding: '0px 12px',
+                            minHeight: 34,
+                            cursor: settings.CTAButton?.CTALinkField ? 'pointer' : 'normal',
+                          }}
+                          onClick={() => {
+                            if (settings.CTAButton?.CTALinkField) {
+                              window.open(`${get(dataItem, settings.CTAButton?.CTALinkField)}`, '_blank');
+                            }
+                          }}
+                        >
+                          <Tooltip
+                            title={get(dataItem, settings.CTAButton?.CTAButtonText || '') || settings.CTAButton?.CTAButtonText || ''}
+                            placement="top"
+                            arrow={true}
+                          >
+                            <Typography
+                              className="text-f16 fw-600 max-line-1"
+                              align="left"
+                              style={{
+                                color: settings.CTAButton?.CTAButtonTextColor || '#FFFFFF',
+                                letterSpacing: '0.55px',
+                                maxWidth: '220px',
+                                paddingRight: '4px',
+                              }}
+                            >
+                              {get(dataItem, settings.CTAButton?.CTAButtonText || '') || settings.CTAButton?.CTAButtonText || ''}
+                            </Typography>
+                          </Tooltip>
+                          {settings.CTAButton?.CTAIcon && (
+                            <Icon
+                              name="link"
+                              fill={settings.CTAButton?.CTAButtonTextColor || '#FFFFFF'}
+                              width={16}
+                            />
+                          )}
+                        </div>
+                      </div>
+                    )}
                   </div>
+                  
                   {productDetails && (
                     <div className="w-100">
                       <Button
@@ -573,10 +539,10 @@ function ProductDetailView(props: Props) {
                           paddingRight: '15px',
                           textTransform: 'initial',
                         }}
-                        onClick={() => setCollapDescription(e => !e)}
+                        onClick={() => setCollapseDescription(e => !e)}
                       >
                         {t('View details')}
-                        {collapDescription ? (
+                        {collapseDescription ? (
                           <KeyboardArrowUpIcon
                             htmlColor={settings.theme?.secondaryColor}
                           />
@@ -586,7 +552,7 @@ function ProductDetailView(props: Props) {
                           />
                         )}
                       </Button>
-                      <Collapse in={collapDescription}>
+                      <Collapse in={collapseDescription}>
                         <Typography
                           style={{
                             fontSize: 14,
@@ -604,94 +570,6 @@ function ProductDetailView(props: Props) {
                 </Grid>
               </Grid>
             </div>
-  
-            {settings.showFeedbackAndShare && (
-              <div
-                className="box-bottom"
-                style={{
-                  height: '48px',
-                  padding: '0px 16px 0px 16px',
-                  marginBottom: 10,
-                  marginTop: 10,
-                  display: 'flex',
-                  justifyContent: 'center',
-                }}
-              >
-                <Grid
-                  container
-                  justifyContent={
-                    settings.shareOption ? 'space-between' : 'space-around'
-                  }
-                  alignItems="center"
-                >
-                  <Grid item>
-                    <div style={{ display: 'flex', alignItems: 'center' }}>
-                      <Button
-                        className="btn-item"
-                        onClick={() => {
-                          handlerFeedback('like');
-                          setFeedback('like');
-                        }}
-                      >
-                        <IconLike
-                          width={24}
-                          height={24}
-                          color={feedback === 'like' ? '#3E36DC' : '#000000'}
-                        />
-                      </Button>
-                    </div>
-                  </Grid>
-                  <Grid item>
-                    <div style={{ display: 'flex', alignItems: 'center' }}>
-                      <Button
-                        className="btn-item"
-                        onClick={() => {
-                          handlerFeedback('dislike');
-                          setFeedback('dislike');
-                        }}
-                      >
-                        <IconDisLike
-                          width={24}
-                          height={24}
-                          color={feedback === 'dislike' ? '#CC1854' : '#000000'}
-                        />
-                      </Button>
-                    </div>
-                  </Grid>
-                  {settings.shareOption && (
-                    <Grid item>
-                      <div style={{ display: 'flex', alignItems: 'center' }}>
-                        <Button
-                          className="btn-item"
-                          onClick={() => onHandlerModalShare()}
-                        >
-                          <IconShare width={24} height={24} color="#000000" />
-                        </Button>
-                      </div>
-                    </Grid>
-                  )}
-                  {/* <Grid item>
-                <div display={'flex'} alignItems={'center'}>
-                  <Button className="btn-item">
-                    <div
-                      className=""
-                      display={'flex'}
-                      justifyContent={'center'}
-                      alignItems={'center'}
-                    >
-                      <img
-                        src={IconSupport}
-                        alt="image_item"
-                        className="icon_support"
-                        style={{ width: '30px' }}
-                      />
-                    </div>
-                  </Button>
-                </div>
-              </Grid> */}
-                </Grid>
-              </div>
-            )}
           </div>
         )}
       </div>