+++ CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.3.61 - 2024-10-24
+
+### Added
+- Pre-filter and metadata CTA link for Widget
+- Field flexibility for Search Suite
+- Feedback for search suite
+
+### Tech debts
+- Move Icons to react-components
+- Remove Cadenas codes and use nyris assets instead
+
 ## 0.3.60 - 2024-10-16
 
 ### Fix